requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
EasyProcess==0.3
idna==2.10
MouseInfo==0.1.3
Pillow==8.2.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-telegram-bot==13.4.1
python3-xlib==0.15
PyTweening==1.0.3
pytz==2021.1
PyVirtualDisplay==2.1
requests==2.25.1
selenium==3.141.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
